+++ src/components/comp-Pago.jsx
@@ -1,77 +1,11 @@
-import React, { useState } from "react";
-import "../assets/styles/comp_pago.css"; // Importa el archivo CSS
-// ResponsiveLayout.jsx
-//import Estado from "./comp-pago-2"
 
+// comp-Pago.jsx
+import React from "react";
+import "../assets/styles/comp_pago.css";
 
-
-const MetodoPago = ({ formData }) => {
-  const [subject, setSubject] = useState("");
-  const [text, setText] = useState("");
-  const [file, setFile] = useState(null);
-  const [metodoPago, setMetodoPago] = useState("");
-
+const MetodoPago = ({ formData, metodoPago, setMetodoPago, handleFileChange, fileChosen }) => {
   const handleMetodoPagoChange = (event) => {
     setMetodoPago(event.target.value);
-  };
-
-  const handleFileChange = (event) => {
-    setFile(event.target.files[0]);
-    document.getElementById("file-chosen").textContent =
-      event.target.files[0].name;
-  };
-
-  const handleClickEnviar = async () => {
-    const formDataToSend = new FormData();
-    formDataToSend.append("to", "programacionprueba99@gmail.com");
-    formDataToSend.append("subject", subject);
-    formDataToSend.append(
-      "text",
-      `Información de el Huesped; PAGO
-        Nombre: ${formData.nombreH}
-        Apellido: ${formData.apellidoH}
-        Teléfono: ${formData.telefonoH}
-        Email: ${formData.emailH}
-        Pago: ${metodoPago}
-        ${formData.vehiculoH ? `Vehículo: Sí` : `Vehículo: No`}
-        ${formData.vehiculoH ? `Tipo de vehículo: ${formData.tipoH}` : ""}
-        ${formData.vehiculoH ? `Marca y modelo: ${formData.marcamodeloH}` : ""}
-        ${formData.vehiculoH ? `Color: ${formData.colorH}` : ""}
-        ${formData.vehiculoH ? `Patente: ${formData.patenteH}` : ""}
-        <p>Mensaje: ${text}
-
-       `
-    );
-    if (file) {
-      formDataToSend.append("attachment", file);
-    }
-
-    try {
-      const response = await fetch(
-        "http://localhost:5000/api/email/send-email",
-        {
-          method: "POST",
-          body: formDataToSend,
-        }
-      );
-
-      if (response.ok) {
-        alert("Correo electrónico enviado con éxito!");
-        setSubject("");
-        setText("");
-        setFile(null);
-        setMetodoPago("");
-        document.getElementById("file-chosen").textContent =
-          "Ningún archivo seleccionado";
-      } else {
-        const errorData = await response.json();
-        console.error("Error al enviar el correo:", errorData);
-        alert("Hubo un error al enviar el correo electrónico.");
-      }
-    } catch (error) {
-      console.error("Error al enviar el correo:", error);
-      alert("Hubo un error al enviar el correo electrónico.");
-    }
   };
 
   return (
@@ -141,7 +75,7 @@ const MetodoPago = ({ formData }) => {
           </p>
         </div>
 
-        <div className=" espacio file-upload">
+        <div className="espacio file-upload">
           <input
             type="file"
             id="upload-file"
@@ -152,35 +86,35 @@ const MetodoPago = ({ formData }) => {
           <label htmlFor="upload-file">
             <span>Subir Comprobante</span>
           </label>
-          <div id="file-chosen">Ningún archivo seleccionado</div>
+          <div id="file-chosen">{fileChosen}</div>
         </div>
       </div>
-      <button className="enviar-button" onClick={handleClickEnviar}>
-        Enviar
-      </button>
     </div>
   );
 };
 
-
-
-const ResponsiveLayout = ({ formData , numberOfRooms , start , end }) => {
-    // Datos de ejemplo para el carrito
+const ResponsiveLayout = ({ formData, numberOfRooms, start, end, metodoPago, setMetodoPago, handleFileChange, fileChosen }) => {
     const itemsCarrito = Array.from({ length: numberOfRooms }, (_, index) => ({ nombre: `Habitación ${index + 1}` }));
-    console.log(itemsCarrito);
-        return (
+
+    return (
         <div className="espacio layout-container">
             <div className="left-section">
-<MetodoPago formData={formData} items={itemsCarrito} /> 
+                <MetodoPago 
+                    formData={formData} 
+                    metodoPago={metodoPago}
+                    setMetodoPago={setMetodoPago}
+                    handleFileChange={handleFileChange}
+                    fileChosen={fileChosen}
+                />
             </div>
 
             <div className="right-section">
                 <div className="top-box">
                     <div className="content-box">
                         <h3>Detalles de la Reserva</h3>
-                        <p>Fecha de Entrada: {start}</p><p>Fecha de Salida: {end}</p>
+                        <p>Fecha de Entrada: {start}</p>
+                        <p>Fecha de Salida: {end}</p>
 
-                        {/* Resumen del Carrito */}
                         <h3>Resumen del Carrito</h3>
                         <ul>
                             {itemsCarrito.map((item, index) => (
@@ -203,5 +137,4 @@ const ResponsiveLayout = ({ formData , numberOfRooms , start , end }) => {
         </div>
     );
 };
-
 export default ResponsiveLayout;
